[wip] feat (stocklet): preparing and caching JWKs
Auth service will prepare a JWK response with the
public key to allow the API ingress to validate JWT
tokens.

The process is a little hacky at the minute (with
conversion from the JWK object -> JSON -> Protobuf).

Hopefully I can implement an improved solution in the
future; but for now IIABDFI.

todo:
- remaining auth service methods
- auth service postgres/storage controllers
- service wrapping (serve gRPC, gRPC-Gateway - consume events)