feat (stocklet-proto1): addition of gRPC status to svc errors
This helps to prevent leaking details from errors being wrapped by the service errors.

The goal is for every error which could be exposed to the user be wrapped in a service error, otherwise fallback to a standard error message (in the top-level error handler).

Additionally, since we can interface directly with the HTTP error interface, there is room to expand further than the limited number of gRPC status codes available.
Better matching HTTP status codes?